Convert one_hot_encodings to float32 for compatibility with mps

+++ utils.py
@@ -10,7 +10,7 @@ def clip_and_normalize_image(array, min_value, max_value):
     return ((array - min_value) / max_value)
 
 def one_hot_encode(mask):
-    one_hot_mask = np.zeros((mask.shape[0], mask.shape[1], 2))
+    one_hot_mask = np.zeros((mask.shape[0], mask.shape[1], 2), dtype=np.float32)
     one_hot_mask[mask == 0, 0] = 1
     one_hot_mask[mask == 1, 1] = 1
     assert(np.sum(one_hot_mask) == mask.shape[0] * mask.shape[1])
